Move behavior from celery_user_rate_limit model class to base_task module.
Moved behavioral logic from CeleryUserRateLimit model class to base_task
module because model classes should be plain old classes. Ran
make config-schema to update galaxy_options.rst and galaxy.yml.sample
based on changes made to config_schema.yml.
