+++ doc/source/admin/galaxy_options.rst
@@ -5137,10 +5137,9 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 :Description:
-    Applies if `celery_user_rate_limit` is non-zero. Used for
-    testing against a postgres db. Forces use of standard sql
-    code rather than postgres specific dialect for database
-    access code.
+    Applies if celery_user_rate_limit is non-zero. Used for testing
+    against a postgres db. Forces use of standard sql code rather than
+    postgres specific dialect for database access code.
 :Default: ``false``
 :Type: bool
