+++ lib/galaxy/app.py
@@ -30,7 +30,8 @@ from galaxy import (
 )
 from galaxy.celery.base_task import (
     GalaxyTaskBeforeStart,
-    GalaxyTaskUserRateLimitBeforeStart,
+    GalaxyTaskBeforeStartUserRateLimitPostgres,
+    GalaxyTaskBeforeStartUserRateLimitStandard,
 )
 from galaxy.config_watchers import ConfigWatchers
 from galaxy.datatypes.registry import Registry
@@ -61,8 +62,6 @@ from galaxy.managers.workflows import (
     WorkflowsManager,
 )
 from galaxy.model import (
-    CeleryUserRateLimitPostgres,
-    CeleryUserRateLimitStandard,
     custom_types,
     mapping,
 )
@@ -595,17 +594,20 @@ class GalaxyManagerApplication(MinimalManagerApp, MinimalGalaxyApplication):
         """
         Register subtype class instance to support implementation of a user rate limit for execution of celery tasks.
         The default supertype class does not enforce a user rate limit. This is the case if the celery_user_rate_limit
-        config param is default value of 0.0
+        config param is the default value.
         """
         if self.config.celery_user_rate_limit:
-            force_standard_before_start = self.config.get("celery_user_rate_limit_standard_before_start", False)
-            if not force_standard_before_start and self.config.database_connection.startswith("postgres"):
-                task_before_start = GalaxyTaskUserRateLimitBeforeStart(
-                    self.config.celery_user_rate_limit, CeleryUserRateLimitPostgres(), self.model.session
+            task_before_start: GalaxyTaskBeforeStart
+            if (
+                not self.config.celery_user_rate_limit_standard_before_start
+                and self.config.database_connection.startswith("postgres")
+            ):
+                task_before_start = GalaxyTaskBeforeStartUserRateLimitPostgres(
+                    self.config.celery_user_rate_limit, self.model.session
                 )
             else:
-                task_before_start = GalaxyTaskUserRateLimitBeforeStart(
-                    self.config.celery_user_rate_limit, CeleryUserRateLimitStandard(), self.model.session
+                task_before_start = GalaxyTaskBeforeStartUserRateLimitStandard(
+                    self.config.celery_user_rate_limit, self.model.session
                 )
             self._register_singleton(GalaxyTaskBeforeStart, task_before_start)
