+++ lib/galaxy/config/sample/galaxy.yml.sample
@@ -2736,12 +2736,12 @@ galaxy:
   # https://docs.galaxyproject.org/en/master/admin/production.html
   #enable_celery_tasks: false
 
-  # If set to a non-0 value, upper limit on number of
-  # tasks that can be executed per user per second.
+  # If set to a non-0 value, upper limit on number of tasks that can be
+  # executed per user per second.
   #celery_user_rate_limit: 0.0
 
-  # Applies if celery_user_rate_limit is non-zero. Used for testing against
-  # a postgres db. Forces use of standard sql code rather than
+  # Applies if celery_user_rate_limit is non-zero. Used for testing
+  # against a postgres db. Forces use of standard sql code rather than
   # postgres specific dialect for database access code.
   #celery_user_rate_limit_standard_before_start: false
